[refactor]: 메인 페이지 모달 컴포넌트 분리 완료 #158
