+++ src/modules/survey/main/components/ClosingSurvey.tsx
@@ -1,30 +1,22 @@
 import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { Swiper, SwiperSlide } from 'swiper/react';
-import { SwiperOptions } from 'swiper/types/swiper-options';
 import FaceIcon from '@mui/icons-material/Face';
 import './Modal.css';
 import {
-  Button,
   Card,
   CardContent,
   Typography,
   Box,
   Chip,
   Stack,
-  Modal,
-  Fade,
-  Divider,
-  Alert,
   CardActionArea,
 } from '@mui/material';
-import Swal from 'sweetalert2';
 import CardMedia from '@mui/material/CardMedia';
-import Avatar from '@mui/material/Avatar';
-import ClearTwoToneIcon from '@mui/icons-material/ClearTwoTone';
 import EventAvailableIcon from '@mui/icons-material/EventAvailable';
 import '../../../../global.css';
 import { CardDataListProps, CardDataProps } from '../types/MainType';
+import MainModal from './MainModal';
+import { swiperParams, tagColor } from '../constant/MainConstant';
 
 const styles = {
   CardSwiper: {
@@ -47,24 +39,10 @@ const textStyle = {
   fontFamily,
   textOverflow: 'ellipsis',
 };
-const modalSubText = {
-  fontSize: '15px',
-  marginBottom: '10px',
-  color: '#858585',
-};
-
-// const titleStyle = {
-//   display: 'flex',
-//   fontFamily,
-//   textOverflow: 'ellipsis',
-//   justifyContent: 'center',
-// };
 
 function ClosingSurvey({ cardList }: CardDataListProps) {
   const [openModal, setOpenModal] = useState(false);
   const [selectedCard, setSelectedCard] = useState<CardDataProps | null>(null);
-
-  const navigate = useNavigate();
 
   const getChipColor = (surveyStatusName: string) => {
     switch (surveyStatusName) {
@@ -76,25 +54,6 @@ function ClosingSurvey({ cardList }: CardDataListProps) {
         return '#D7D3D3';
     }
   };
-  const showSwalAlert = () => {
-    Swal.fire({
-      icon: 'warning',
-      title: '설문 참여자가 아직 없습니다.',
-      customClass: {
-        popup: 'swal-custom-popup',
-        container: 'swal-custom-container',
-      },
-    });
-  };
-
-  const numUser = () => {
-    const loginUserNo = localStorage.getItem('userNo');
-    const numUserNo =
-      loginUserNo !== null && loginUserNo !== undefined
-        ? Number(loginUserNo)
-        : null;
-    return numUserNo;
-  };
 
   const openCardModal = (card: CardDataProps) => {
     setSelectedCard(card);
@@ -104,55 +63,6 @@ function ClosingSurvey({ cardList }: CardDataListProps) {
   const closeCardModal = () => {
     setSelectedCard(null);
     setOpenModal(false);
-  };
-
-  const swiperParams: SwiperOptions = {
-    slidesPerView: 'auto',
-    breakpoints: {
-      920: {
-        slidesPerView: 4,
-        spaceBetween: 10,
-      },
-      750: {
-        slidesPerView: 4,
-        spaceBetween: 5,
-      },
-
-      540: {
-        slidesPerView: 3,
-        spaceBetween: 5,
-      },
-
-      500: {
-        slidesPerView: 2,
-        spaceBetween: 5,
-      },
-
-      0: {
-        slidesPerView: 1.8,
-        spaceBetween: 5,
-      },
-    },
-  };
-
-  const tagColor = (tag: string) => {
-    switch (tag) {
-      case '공지':
-        return '#F8E5E5';
-      case '중요':
-        return '#F5F9DE';
-      case '업무':
-        return '#F9ECDF';
-      case '기타':
-        return '#E5ECF5';
-      case '일상':
-        return '#EDEBF6';
-      default:
-        return 'default';
-    }
-  };
-  const handleIconClick = () => {
-    closeCardModal();
   };
 
   return (
@@ -209,20 +119,6 @@ function ClosingSurvey({ cardList }: CardDataListProps) {
                             image={card.surveyImage}
                             alt="survey image"
                           />
-                          {/* <Card
-                        variant="elevation"
-                        sx={{
-                          width: '156px',
-                          height: '180px',
-                          marginLeft: '5px',
-                          borderRadius: 2,
-                          backgroundColor: '#F2F2F2',
-                          boxShadow: '0px 4px 8px rgba(0, 0, 0, 0.1)',
-                        }}
-                        style={textStyle}
-                        onClick={() => openCardModal(card)}
-                        role="button"
-                      > */}
                           <CardContent
                             sx={{
                               padding: '8px',
@@ -350,318 +246,13 @@ function ClosingSurvey({ cardList }: CardDataListProps) {
           </Swiper>
         </Box>
 
-        <Modal
-          open={openModal}
-          onClose={closeCardModal}
-          aria-labelledby="modal-title"
-          aria-describedby="modal-description"
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-          }}
-        >
-          <Fade in={openModal}>
-            <div className="modal">
-              <Box>
-                <Box
-                  sx={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'space-between',
-                  }}
-                >
-                  <div>
-                    <Chip
-                      key="0"
-                      label={selectedCard?.openStatusName}
-                      size="small"
-                      style={textStyle}
-                      sx={{
-                        fontSize: 14,
-                        marginRight: 1,
-                        height: '25px',
-                        backgroundColor: tagColor('0'),
-                        opacity: 0.7,
-                      }}
-                    />
-                    {numUser() === selectedCard?.userNo && (
-                      <Chip
-                        label="본인 작성"
-                        size="small"
-                        style={textStyle}
-                        sx={{
-                          fontSize: 14,
-                          marginRight: 1,
-                          height: '25px',
-                          backgroundColor: tagColor('0'),
-                          opacity: 0.7,
-                        }}
-                      />
-                    )}
-                  </div>
-                  <ClearTwoToneIcon
-                    onClick={handleIconClick}
-                    sx={{
-                      '&:hover': {
-                        cursor: 'pointer',
-                      },
-                    }}
-                  />
-
-                  {/* 닫기 아이콘 */}
-                </Box>
-                <Box
-                  sx={{
-                    display: 'flex',
-                    justifyContent: 'flex-start',
-                    alignItems: 'center',
-                  }}
-                >
-                  <Avatar
-                    src={selectedCard?.userImage}
-                    sx={{
-                      width: 28,
-                      height: 28,
-                      marginRight: '8px',
-                    }}
-                  />
-                  <Typography
-                    sx={{
-                      display: 'flex',
-                      alignItems: 'center',
-                      color: '#393939',
-
-                      height: '45px',
-                      fontWeight: '700',
-                    }}
-                  >
-                    {selectedCard ? selectedCard.userNickName : ''}
-                  </Typography>
-                </Box>
-                {/* 설문 조사 타이틀 */}
-                <Box
-                  className="titleStyle"
-                  style={{
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'flex-start',
-                    height: '80px',
-                  }}
-                >
-                  <Typography
-                    variant="h6"
-                    id="modal-title"
-                    style={{
-                      fontFamily,
-                      textOverflow: 'ellipsis',
-                      fontWeight: 'bold',
-                      marginBottom: '15px',
-                    }}
-                  >
-                    {selectedCard ? selectedCard.surveyTitle : ''}
-                  </Typography>
-                </Box>
-
-                {/* 작성자, 참여자수, 태그들 */}
-                <Box
-                  sx={{
-                    display: 'flex',
-                    justifyContent: 'space-between',
-                    flexDirection: 'column',
-                  }}
-                >
-                  {/* 설문 조사 기간 */}
-                  <Typography
-                    style={modalSubText}
-                    sx={{ display: 'flex', alignItems: 'center' }}
-                  >
-                    <EventAvailableIcon
-                      sx={{
-                        fontSize: '18px',
-                        marginRight: '4px',
-                      }}
-                    />{' '}
-                    {selectedCard
-                      ? ` ${selectedCard.surveyPostAt.slice(0, 10)} ~ ${
-                          selectedCard.surveyClosingAt
-                        }`
-                      : ''}
-                  </Typography>
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      justifyContent: 'space-between',
-                      flexDirection: 'row',
-                    }}
-                  >
-                    <Typography
-                      sx={{
-                        color: '#808080',
-                        display: 'flex',
-                        justifyContent: 'flex-start',
-                        alignItems: 'center',
-                      }}
-                    >
-                      <FaceIcon
-                        sx={{
-                          fontSize: '20px',
-                          marginRight: '8px',
-                        }}
-                      />{' '}
-                      {selectedCard ? selectedCard.surveyAttendCount : ''}
-                    </Typography>
-                    <Box sx={{ display: 'flex', justifyContent: 'flex-end' }}>
-                      {selectedCard?.tagName.map((tag) => (
-                        <Chip
-                          key={tag}
-                          label={tag}
-                          size="small"
-                          style={textStyle}
-                          sx={{
-                            fontSize: 12,
-                            marginRight: 1,
-                            height: '25px',
-                            backgroundColor: tagColor(tag),
-                            opacity: 0.7,
-                          }}
-                        />
-                      ))}
-                    </Box>
-                  </Box>
-                </Box>
-                <Divider sx={{ marginBottom: '10px', marginTop: '10px' }} />
-
-                <Box
-                  className="modal-scroll-box"
-                  sx={{
-                    overflow: 'auto',
-                    height: '28vh',
-                    '@media screen and (min-height: 1000px)': {
-                      overflow: 'auto',
-                      height: '36vh',
-                    },
-
-                    '@media screen and (min-width: 374px) and (max-width: 600px) and (min-height: 800px) and (max-height: 1000px)':
-                      {
-                        height: '23vh',
-                      },
-
-                    '@media screen and  (max-width: 376px) and (max-width: 600px)':
-                      {
-                        height: '24vh',
-                      },
-                    '@media screen and (min-width: 374px) and (max-width: 600px) and (min-height: 500px) and (max-height: 700px)':
-                      {
-                        overflow: 'auto',
-
-                        height: '20vh',
-                      },
-                  }}
-                >
-                  {/* 설문조사 사진 */}
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      flexDirection: 'row',
-                      paddingBottom: '15px',
-                    }}
-                  >
-                    <img
-                      src={selectedCard?.surveyImage}
-                      alt="Survey"
-                      style={{ width: '100%', height: 'auto' }}
-                    />{' '}
-                  </Box>
-
-                  <Typography
-                    id="modal-description"
-                    style={{
-                      display: 'flex',
-                      justifyContent: 'flex-start',
-                      alignItems: 'center',
-                      textAlign: 'start',
-                      paddingBottom: '15px',
-                      fontFamily,
-                    }}
-                  >
-                    {selectedCard
-                      ? `설문 설명: ${selectedCard.surveyDescription}`
-                      : ''}
-                  </Typography>
-                </Box>
-                <Divider sx={{ marginBottom: '10px', marginTop: '10px' }} />
-
-                <Box
-                  sx={{
-                    marginTop: '15px',
-                    paddingBottom: '15px',
-                  }}
-                >
-                  {selectedCard?.openStatusName === '비공개' &&
-                    (numUser() !== selectedCard.userNo ? (
-                      <Alert severity="warning">
-                        설문 작성자만 볼 수 있습니다.
-                      </Alert>
-                    ) : (
-                      <Alert severity="success">
-                        해당 비공개 설문의 작성자입니다.
-                      </Alert>
-                    ))}
-
-                  {selectedCard?.openStatusName === '회원 공개' &&
-                    numUser() === null && (
-                      <Alert severity="error">
-                        결과를 보시려면 로그인해주세요.
-                      </Alert>
-                    )}
-                </Box>
-                {/* 결과보기, 참여하기 버튼 */}
-                {/* 결과보기 버튼 */}
-                {(!selectedCard?.openStatusName ||
-                  selectedCard?.openStatusName === '전체 공개' ||
-                  (selectedCard?.openStatusName === '비공개' &&
-                    numUser() !== null &&
-                    numUser() === selectedCard?.userNo) ||
-                  (selectedCard?.openStatusName === '회원 공개' &&
-                    numUser() !== null)) && (
-                  <Button
-                    onClick={() => {
-                      if (selectedCard?.surveyAttendCount === 0) {
-                        showSwalAlert();
-                      } else {
-                        navigate(
-                          `/survey/statistics/${selectedCard?.surveyNo}`
-                        );
-                      }
-                    }}
-                    sx={{
-                      width: '100%',
-                      marginBottom: '8px',
-                      backgroundColor: '#ebebeb',
-                      '&:hover': {
-                        backgroundColor: 'gray',
-                        color: 'white',
-                        fontWeight: '900',
-                        fontSize: '15px',
-                      },
-                      color: 'black',
-                      fontWeight: '600',
-                      '@media (min-height: 801px) and (max-height: 1000px)': {
-                        marginTop: '90px',
-                      },
-                      '@media (max-width: 375px)': {
-                        marginTop: '50px',
-                      },
-                    }}
-                  >
-                    설문 결과보기
-                  </Button>
-                )}
-              </Box>
-            </div>
-          </Fade>
-        </Modal>
+        {openModal && selectedCard && (
+          <MainModal
+            openModal={openModal}
+            closeCardModal={closeCardModal}
+            selectedCard={selectedCard}
+          />
+        )}
       </div>
     </div>
   );
